2898 Replace `Environment` with `environmentId` across controllers, services, and mappers.

+++ server/ee/apps/runtime-job-app/src/main/java/com/bytechef/runtime/job/platform/connection/service/ConnectionServiceImpl.java
@@ -64,7 +64,7 @@ public class ConnectionServiceImpl implements ConnectionService {
 
     @Override
     public List<Connection> getConnections(
-        String componentName, Integer connectionVersion, Long tagId, Integer environmentId, ModeType type) {
+        String componentName, Integer connectionVersion, Long tagId, Long environmentId, ModeType type) {
 
         throw new UnsupportedOperationException();
     }
