+++ server/ee/libs/platform/platform-connection/platform-connection-remote-client/src/main/java/com/bytechef/ee/platform/connection/remote/client/fasade/RemoteConnectionFacadeClient.java
@@ -44,7 +44,7 @@ public class RemoteConnectionFacadeClient implements ConnectionFacade {
 
     @Override
     public List<ConnectionDTO> getConnections(
-        String componentName, Integer connectionVersion, List<Long> connectionIds, Long tagId, Integer environmentId,
+        String componentName, Integer connectionVersion, List<Long> connectionIds, Long tagId, Long environmentId,
         ModeType type) {
 
         throw new UnsupportedOperationException();
